app/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
azure-storage-blob==12.24.0
gunicorn==23.0.0
python-dotenv==1.0.1
